+++ engine/src/engine/Method.java
@@ -4,6 +4,6 @@ import java.io.Serializable;
 
 public interface Method extends Serializable {
     int getValue();
-    float calculateOptionValue(int firstOptionShares, int secondOptionShares);
-    float calculateBalance(int totalSharesBought, int totalSharesBought1);
+    double calculateOptionValue(int firstOptionShares, int secondOptionShares);
+    double calculateBalance(int totalSharesBought, int totalSharesBought1);
 }
